Added OpenIOC and MISP xml parsing for easier automated attribute population
So far, populating attributes through the API is a manual process, with the only exception being an out of date example demonstrating how to map IOC xml attributes to MISP xml attributes. MISP's web UI has an interface to allow users to upload their OpenIOC xml files for importing, but this same functionality is not replicated in the PyMISP apis.

I needed this functionality, so I tested the examples, and then tested the new apis after I added them to the PyMISP api.
For each IOC or MISP xml api offered, the user can either pass in a filename or a BytesIO object that can be manipulated with read. This prevents the user who has the xml information in memory from having to needlessly save it to disk in the case that an open call is made within the api by splitting the interpretation of the xml from the input type requirements.